+++ smac/facade/experimental/hydra_facade.py
@@ -19,8 +19,7 @@ from smac.scenario.scenario import Scenario
 from smac.facade.smac_ac_facade import SMAC4AC
 from smac.facade.experimental.psmac_facade import PSMAC
 from smac.utils.io.output_directory import create_output_directory
-from smac.runhistory.runhistory import RunHistory
-from smac.optimizer.objective import average_cost
+from smac.runhistory.runhistory import RunHistory, average_cost
 from smac.epm.util_funcs import get_rng
 from smac.utils.constants import MAXINT
 from smac.optimizer.pSMAC import read
