+++ smac/intensification/abstract_racer.py
@@ -5,13 +5,12 @@ from collections import OrderedDict
 
 import numpy as np
 
-from smac.optimizer.objective import sum_cost
 from smac.optimizer.epm_configuration_chooser import EPMChooser
 
 from smac.stats.stats import Stats
 from smac.utils.constants import MAXINT
 from smac.configspace import Configuration
-from smac.runhistory.runhistory import RunHistory
+from smac.runhistory.runhistory import RunHistory, sum_cost
 from smac.tae.execute_ta_run import ExecuteTARun
 from smac.utils.io.traj_logging import TrajLogger
