+++ smac/smac_cli.py
@@ -8,14 +8,12 @@ from smac.utils.io.cmd_reader import CMDReader
 from smac.scenario.scenario import Scenario
 from smac.facade.smac_ac_facade import SMAC4AC
 from smac.facade.roar_facade import ROAR
-from smac.facade.experimental.epils_facade import EPILS
 from smac.facade.experimental.hydra_facade import Hydra
 from smac.facade.experimental.psmac_facade import PSMAC
 from smac.facade.smac_hpo_facade import SMAC4HPO
 from smac.facade.smac_bo_facade import SMAC4BO
-from smac.runhistory.runhistory import RunHistory
+from smac.runhistory.runhistory import RunHistory, average_cost
 from smac.stats.stats import Stats
-from smac.optimizer.objective import average_cost
 from smac.utils.merge_foreign_data import merge_foreign_data_from_file
 from smac.utils.io.traj_logging import TrajLogger
 from smac.tae.execute_ta_run import TAEAbortException, FirstRunCrashedException
@@ -136,13 +134,6 @@ class SMACCLI(object):
                 run_id=main_args_.seed)
         elif main_args_.mode == "ROAR":
             optimizer = ROAR(
-                scenario=scen,
-                rng=np.random.RandomState(main_args_.seed),
-                runhistory=rh,
-                initial_configurations=initial_configs,
-                run_id=main_args_.seed)
-        elif main_args_.mode == "EPILS":
-            optimizer = EPILS(
                 scenario=scen,
                 rng=np.random.RandomState(main_args_.seed),
                 runhistory=rh,
