+++ smac/utils/validate.py
@@ -11,8 +11,7 @@ from smac.configspace import Configuration, convert_configurations_to_array
 from smac.epm.rf_with_instances import RandomForestWithInstances
 from smac.epm.rfr_imputator import RFRImputator
 from smac.epm.util_funcs import get_types
-from smac.optimizer.objective import average_cost
-from smac.runhistory.runhistory import RunHistory, RunKey, StatusType
+from smac.runhistory.runhistory import RunHistory, RunKey, StatusType, average_cost
 from smac.runhistory.runhistory2epm import RunHistory2EPM4Cost
 from smac.scenario.scenario import Scenario
 from smac.stats.stats import Stats
