+++ test/test_intensify/test_successive_halving.py
@@ -10,8 +10,7 @@ from ConfigSpace.hyperparameters import UniformIntegerHyperparameter
 from smac.scenario.scenario import Scenario
 from smac.tae.execute_func import ExecuteTAFuncDict
 from smac.intensification.successive_halving import SuccessiveHalving
-from smac.runhistory.runhistory import RunHistory
-from smac.optimizer.objective import average_cost
+from smac.runhistory.runhistory import RunHistory, average_cost
 from smac.tae.execute_ta_run import StatusType
 from smac.stats.stats import Stats
 from smac.utils.io.traj_logging import TrajLogger
