+++ test/test_runhistory/test_runhistory.py
@@ -7,8 +7,7 @@ from ConfigSpace import Configuration, ConfigurationSpace
 from ConfigSpace.hyperparameters import UniformIntegerHyperparameter
 
 from smac.tae.execute_ta_run import StatusType
-from smac.runhistory.runhistory import RunHistory
-from smac.optimizer.objective import average_cost
+from smac.runhistory.runhistory import RunHistory, average_cost
 
 
 def get_config_space():
